+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Current Master
 
+- Add ability to import accounts in JSON file format (used by Mist, Geth, MyEtherWallet, and more!)
+
 ## 3.1.0 2017-1-18
 
 - Add ability to import accounts by private key.